# docs/swagger.yaml
basePath: /
definitions:
...
  api.liteLLMCreditAdjustRequest:
    properties:
      amount:
        type: number
      mode:
        type: string
      reason:
        type: string
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
...
  api.swaggerAPIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.swaggerErrorBody'
    type: object
...
  api.swaggerAuditListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.AuditEvent'
        type: array
    type: object
...
  api.swaggerLiteLLMAccessResponse:
    properties:
      can_manage:
        type: boolean
      service_available:
        type: boolean
    type: object
  api.swaggerLiteLLMCallListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/litellmcredit.RecentCall'
        type: array
      limit:
        type: integer
    type: object
...
  api.swaggerSimpleErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  litellmcredit.CreditSnapshot:
    properties:
      api_key:
        type: string
      budget_remaining:
        type: number
      budget_total:
        type: number
      key_alias:
        type: string
      last_synced_at:
        type: string
      litellm_user_id:
        type: string
      spend_used:
        type: number
      tenant_id:
        type: string
      unit:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the Portal backend.
  title: Portal API
  version: "1.0"
paths:
  /api/v1/admin/litellm/access:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerLiteLLMAccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get LiteLLM admin access flags
      tags:
      - litellm-admin
  /api/v1/admin/litellm/calls/{tenant_id}/{user_id}:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Max items (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerLiteLLMCallListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: List user recent calls (admin)
      tags:
      - litellm-admin
  /api/v1/admin/litellm/credits/{tenant_id}/{user_id}:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/litellmcredit.CreditSnapshot'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user credit snapshot (admin)
      tags:
      - litellm-admin
  /api/v1/admin/litellm/credits/adjust:
    post:
      consumes:
      - application/json
      parameters:
      - description: Adjust payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.liteLLMCreditAdjustRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/litellmcredit.CreditSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Adjust user credit (admin)
      tags:
      - litellm-admin
  /api/v1/admin/litellm/events:
    get:
      parameters:
      - description: Max items (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerLiteLLMEventListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: List credit events (admin)
      tags:
      - litellm-admin
  /api/v1/audit/events:
    get:
      parameters:
      - description: Actor filter
        in: query
        name: actor
        type: string
      - description: Action filter
        in: query
        name: action
        type: string
      - description: Resource filter
        in: query
        name: resource
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerAuditListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Query audit events
      tags:
      - audit
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authn.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerOKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Logout by refresh token
      tags:
      - auth
  /api/v1/auth/password/change:
    post:
      consumes:
      - application/json
      parameters:
      - description: Change password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerOKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      security:
      - BearerAuth: []
      summary: Change current password
      tags:
      - auth
  /api/v1/auth/password/forgot:
    post:
      consumes:
      - application/json
      parameters:
      - description: Forgot password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.forgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerOKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /api/v1/auth/password/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.swaggerOKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Reset password
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authn.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Refresh token pair
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/authn.RegisterResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Register by email
      tags:
      - auth
  /api/v1/auth/register/phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register by phone payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.registerByPhoneRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/authn.RegisterResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerAPIErrorResponse'
      summary: Register by phone
      tags:
      - auth
...
  /api/v1/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/identity.Principal'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerSimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current principal
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: 'Use JWT bearer token: `Bearer <token>`.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"